+++ src/components/SearchBar.jsx
@@ -1,25 +1,17 @@
 import debounce from "@/utils/debounce";
 import { useState } from "react";
 import { useId } from "react";
+import SearchButton from "./SearchButton";
 
 function SearchBar() {
 
   const searchId = useId();
-  const [userInput, setUserInput] = useState();
+  const [searchInput, setUserInput] = useState();
   
   //# 검색창
   const handleChangeInput = debounce((e) => {
-    // console.log(e.target.value);
     setUserInput(e.target.value.toLowerCase())
   }, 300);
-
-  // console.log(userInput);
-
-  //# 검색버튼
-  const handleSearch = () => {
-
-    console.log('검색');
-  }
 
   return (
     <>
@@ -29,18 +21,11 @@ function SearchBar() {
           type="text"
           name="search"
           id={searchId}
-          defaultValue={userInput}
+          defaultValue={searchInput}
           onChange={handleChangeInput}
           className="w-[993px] h-[57px] rounded-[10px] border-2 border-solid border-lionBlack bg-lionWhite mr-2 outline-lionYellow"
         />
-        <div 
-          role="button" 
-          tabIndex={0}
-          aria-labelledby="search"
-          onClick={handleSearch}
-          className="outline-none">
-          <img src="/search.svg" alt="검색" id="search"/>
-        </div>
+      <SearchButton searchInput={searchInput}/>
       </form>
     </>
   )
